restructure views and file paths
